requirements.txt
Pygments==1.6
configparser==3.3.0r2
logging==0.4.9.6
pudb==2013.3.5
pyjade==2.0.2
tornado==3.1
urwid==1.1.1
wsgiref==0.1.2
RPi.GPIO==0.5.3a
RPIO==0.10.0

